+++ x/rollapp/keeper/msg_server_update_state_test.go
@@ -1,9 +1,9 @@
 package keeper_test
 
 import (
+	abci "github.com/cometbft/cometbft/abci/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
-	abci "github.com/tendermint/tendermint/abci/types"
 
 	"github.com/dymensionxyz/dymension/v3/testutil/sample"
 	common "github.com/dymensionxyz/dymension/v3/x/common/types"
@@ -30,6 +30,8 @@ func (suite *RollappTestSuite) TestFirstUpdateState() {
 			Description:  "Sample description",
 			LogoDataUri:  "data:image/png;base64,c2lzZQ==",
 			TokenLogoUri: "data:image/png;base64,ZHVwZQ==",
+			Telegram:     "rolly",
+			X:            "rolly",
 		},
 	}
 	suite.App.RollappKeeper.SetRollapp(suite.Ctx, rollapp)