+++ x/sequencer/keeper/msg_server_create_sequencer_test.go
@@ -122,6 +122,8 @@ func (suite *SequencerTestSuite) TestCreateSequencer() {
 				Description:  "Sample description",
 				LogoDataUri:  "data:image/png;base64,c2lzZQ==",
 				TokenLogoUri: "data:image/png;base64,ZHVwZQ==",
+				Telegram:     "rolly",
+				X:            "rolly",
 			},
 		}
 		suite.App.RollappKeeper.SetRollapp(suite.Ctx, rollapp)