+++ x/vfc/hooks/hooks_test.go
@@ -3,9 +3,9 @@ package hooks_test
 import (
 	"testing"
 
+	cometbftproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	"github.com/dymensionxyz/dymension/v3/app/apptesting"
 	"github.com/stretchr/testify/suite"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 )
 
 type HooksTestSuite struct {
@@ -18,7 +18,7 @@ func TestHooksTestSuite(t *testing.T) {
 
 func (suite *HooksTestSuite) SetupTest() {
 	app := apptesting.Setup(suite.T(), false)
-	ctx := app.GetBaseApp().NewContext(false, tmproto.Header{})
+	ctx := app.GetBaseApp().NewContext(false, cometbftproto.Header{})
 
 	suite.App = app
 	suite.Ctx = ctx